changed the type of variables from int to double

+++ src/calculatorMain.java
@@ -3,18 +3,18 @@ import java.util.Scanner;
 public class calculatorMain {
 
 	public static void main(String[] args) {
-		int number1;
-		int number2;
+		double number1;
+		double number2;
 		char operator;
 		double answer=0.0;
 		
 		Scanner obj = new Scanner (System.in);
 		System.out.println("Please enter first number: ");
-		number1 = obj.nextInt();
+		number1 = obj.nextDouble();
 		System.out.println("write Arithmatic operator + or - or / or * ");
 		operator = obj.next().charAt(0);
 		System.out.println("Please enter second number: ");
-		number2 = obj.nextInt();
+		number2 = obj.nextDouble();
 		
 		switch (operator) {
 		case '+': answer = number1+number2;
